# AK+LibGfx: Use SIMD for PNG score calculation